FIX: Cast y to correct type in validation dataset, quickfix

+++ src/unlearners/scrub.py
@@ -61,7 +61,8 @@ class ScrubR(BaseUnlearner):
         X_values = torch.cat(X_values)
         y_values = torch.cat(y_values)
 
-        dataset = SyntheticDataset(X_values.numpy(), y_values.numpy(), n_classes=n_classes)
+        # Ensure we're using Python native types, not torch.int64
+        dataset = SyntheticDataset(X_values.numpy(), y_values.numpy(), n_classes=int(n_classes))
         dataloader = DataLoader(dataset, batch_size=val_dataloader.batch_size)
         return dataloader
 
@@ -72,6 +73,8 @@ class ScrubR(BaseUnlearner):
         with torch.no_grad():
             for batch in dataloader:
                 x, y = batch[0], batch[1]
+                # convert to same type float32
+                y = y.to(torch.float32)
                 logits = model(x)['logits']
                 loss = self.CE(logits, y)
                 losses.append(loss)
